// src/app/shared/models/employeeModel/edit-view-details.model.ts
export interface IuserInuserDetails {
    FirstName: string;
    LastName: string;
    MobileNo: number;
    UserId: string;
}
export interface IshowData {
    UserType: string;
    FirstName: string;
    LastName: string;
    ProjectName: string;
    FlatNo: number;
    BuildingType: string;
    MobileNo: number;
    Email: string;
    Occupation: string;
    Type: string;
    Organization: string;
    ReraNo: number;
    GSTIN: number;
    AadharNo: string;
    PanNo: string;
}